LeahHow_Goodbye,Earth!
Welcome to Goodbye, Earth! The year is 2021. Due to global warming, the earth is dying. We all knew this was coming but the only person that planned it well was Elon. Yes, Elon Musk. The billionaire that sold flamethrowers and Teslas.
You can either stay on earth and die with it or move to Mars on Elon's rocket.